requirements.txt
Flask
gunicorn
requests
watson_developer_cloud
python-dotenv
virtualenv